#515: Combined Storage integration tests given the max tx size change
